rename: PostCollectionEnableLoader admin 도메인으로 이동
